Modified View for adding new farms

+++ FarmRecordManagementSystem/FarmRecordManagementSystem/Models/Crops.cs
@@ -10,7 +10,7 @@ namespace FarmRecordManagementSystem.Models
         public decimal FarmSizePlanted { get; set; }
         public decimal QuantityPlanted { get; set; }
         public decimal ExpectedHarvestedQuantity { get; set; }
-        public decimal ExpectedBagsHarvested { get; set; }
+        public int ExpectedBagsHarvested { get; set; }
         public DateTime ExpectedHarvestDate { get; set; }
         public DateTime PlantingDate { get; set; }
     }
